Rearrange the chrono project, make the windows chrono stuff a separate target
