+++ src/libraries/chrono/Platforms/Windows/src/windows_chrono_casts.cpp
@@ -1,0 +1,4 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+#include <m/windows_chrono/windows_chrono_casts.h>
