+++ src/libraries/chrono/include/CMakeLists.txt
@@ -4,6 +4,4 @@ target_sources(m_chrono PUBLIC FILE_SET HEADERS FILES
     m/chrono/chrono.h
 )
 
-add_subdirectory(Platforms)
-
 set(m_installation_targets ${m_installation_targets} PARENT_SCOPE)
