+++ src/libraries/chrono/src/CMakeLists.txt
@@ -7,6 +7,5 @@ target_sources(m_chrono PRIVATE
 target_link_libraries(m_chrono PUBLIC
     m_wrappers
 )
-add_subdirectory(Platforms)
 
 set(m_installation_targets ${m_installation_targets} PARENT_SCOPE)
