--- src/libraries/chrono/src/Platforms/Windows/windows_chrono_casts.cpp
@@ -1,4 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT License.
-
-#include <m/chrono/windows_chrono_casts.h>
