+++ src/libraries/chrono/test/CMakeLists.txt
@@ -1,5 +1,3 @@
 cmake_minimum_required(VERSION 3.23)
 
-add_subdirectory(Platforms)
-
 set(m_installation_targets ${m_installation_targets} PARENT_SCOPE)
